+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.6.35] 2026-08-25
+## [0.6.35] 2026-08-27
 
 ### Added
 
 - The API of **IWindowSystem** was extended with a new getter IsRunning().
+
+- The APIs of **IWorld**, **ISystemManager** were extended with new public methods FillFramePacket. A new interface **IRenderSystem** was added to
+mark up systems that require send their data into the renderer's thread.
 
 ### Changed